// apps/client-desktop/src/renderer/src/types.d.ts
export type PlayInfo = {
    code: number
    message: string
    ttl: number
    data: {
        from: string
        result: string
        message: string
        quality: number
        format: string
        timelength: number
        accept_format: string
        accept_description: Array<string>
        accept_quality: Array<number>
        video_codecid: number
        seek_param: string
        seek_type: string
        dash: {
            duration: number
            minBufferTime: number
            min_buffer_time: number
            video: Array<{
                id: number
                baseUrl: string
                base_url: string
                backupUrl: Array<string>
                backup_url: Array<string>
                bandwidth: number
                mimeType: string
                mime_type: string
                codecs: string
                width: number
                height: number
                frameRate: string
                frame_rate: string
                sar: string
                startWithSap: number
                start_with_sap: number
                SegmentBase: {
                    Initialization: string
                    indexRange: string
                }
                segment_base: {
                    initialization: string
                    index_range: string
                }
                codecid: number
            }>
            audio: Array<{
                id: number
                baseUrl: string
                base_url: string
                backupUrl: Array<string>
                backup_url: Array<string>
                bandwidth: number
                mimeType: string
                mime_type: string
                codecs: string
                width: number
                height: number
                frameRate: string
                frame_rate: string
                sar: string
                startWithSap: number
                start_with_sap: number
                SegmentBase: {
                    Initialization: string
                    indexRange: string
                }
                segment_base: {
                    initialization: string
                    index_range: string
                }
                codecid: number
            }>
            dolby: {
                type: number
                audio: any
            }
            flac: any
        }
        support_formats: Array<{
            quality: number
            format: string
            new_description: string
            display_desc: string
            superscript: string
            codecs: Array<string>
        }>
        high_format: any
        volume: {
            measured_i: number
            measured_lra: number
            measured_tp: number
            measured_threshold: number
            target_offset: number
            target_i: number
            target_tp: number
        }
        last_play_time: number
        last_play_cid: number
    }
    session: string
};